Refactor transformer validation
