+++ dita_xml_parser/transformer.py
@@ -6,7 +6,7 @@ import datetime
 import json
 import logging
 import os
-from typing import List, Tuple
+from typing import Dict, List, Tuple
 
 
 from lxml import etree
@@ -14,14 +14,9 @@ from lxml import etree
 import config
 from . import minimal
 from . import utils
+from .validator import DitaValidator, ValidationReport
 
 
-class ValidationReport:
-    """Simple container for validation results."""
-
-    def __init__(self, passed: bool, details: List[str]) -> None:
-        self.passed = passed
-        self.details = details
 
 
 class Dita2LLM:
@@ -49,6 +44,7 @@ class Dita2LLM:
         os.makedirs(self.log_dir, exist_ok=True)
         self.logger = logging.getLogger("Dita2LLM")
         self.logger.setLevel(getattr(logging, config.LOG_LEVEL))
+        self.validator = DitaValidator(self.logger)
         # Path to the most recently generated target file
         self._last_target_path: str | None = None
         self._log_path: str | None = None
@@ -175,7 +171,7 @@ class Dita2LLM:
                 seg_id, value = next(iter(entry.items()))
             elems = root.xpath(f"//*[@data-dita-seg-id='{seg_id}']")
             if not elems:
-                self.logger.error(f"ID {seg_id} not found in skeleton")
+                self.logger.error("ID %s not found in skeleton", seg_id)
                 continue
             elem = elems[0]
             utils.set_inner_xml(elem, value)
@@ -196,8 +192,8 @@ class Dita2LLM:
             pretty_print=True,
         )
         self._last_target_path = target_path
-        self.logger.info(f"Skeleton used: {skeleton_path}")
-        self.logger.info(f"Wrote integrated file: {target_path}")
+        self.logger.info("Skeleton used: %s", skeleton_path)
+        self.logger.info("Wrote integrated file: %s", target_path)
         self.logger.info("End integrate")
         return target_path
 
@@ -206,69 +202,8 @@ class Dita2LLM:
 
         src_xml = self._resolve(src_xml, self.source_dir)
         tgt_xml = self._resolve(tgt_xml, self.target_dir)
-        self.logger.info("Start validate: %s vs %s", src_xml, tgt_xml)
-        parser = etree.XMLParser(remove_blank_text=False)
-        try:
-            src_tree = etree.parse(src_xml, parser)
-            tgt_tree = etree.parse(tgt_xml, parser)
-        except Exception as e:
-            self.logger.error(f"Validation parse error: {e}")
-            return ValidationReport(False, [str(e)])
-        errors = []
-        warnings: List[str] = []
-        if src_tree.docinfo.doctype != tgt_tree.docinfo.doctype:
-            errors.append("DOCTYPE changed")
-
-        def walk(e1, e2, path=""):
-            if isinstance(e1.tag, str) and isinstance(e2.tag, str):
-                if e1.tag != e2.tag:
-                    errors.append(f"tag mismatch at {path}/{e1.tag}")
-                    return
-                if e1.attrib != e2.attrib:
-                    errors.append(f"attrib mismatch at {path}/{e1.tag}")
-                children1 = list(e1)
-                children2 = list(e2)
-                if len(children1) != len(children2):
-                    errors.append(f"child count mismatch at {path}/{e1.tag}")
-                for c1, c2 in zip(children1, children2):
-                    walk(c1, c2, path + "/" + e1.tag)
-            elif isinstance(e1, etree._Comment) and isinstance(e2, etree._Comment):
-                if e1.text != e2.text:
-                    errors.append(f"comment mismatch at {path}")
-            elif isinstance(e1, etree._ProcessingInstruction) and isinstance(
-                e2, etree._ProcessingInstruction
-            ):
-                if e1.target != e2.target or e1.text != e2.text:
-                    errors.append(f"pi mismatch at {path}")
-
-        walk(src_tree.getroot(), tgt_tree.getroot())
-
-        base = os.path.splitext(os.path.basename(src_xml))[0]
         skel_dir = self.intermediate_dir or os.path.dirname(tgt_xml)
-        skeleton_path = os.path.join(skel_dir, f"{base}.skeleton.xml")
-        if os.path.exists(skeleton_path):
-            skel_tree = etree.parse(skeleton_path, parser)
-            skel_et = etree.ElementTree(skel_tree.getroot())
-            for seg_elem in skel_tree.xpath("//*[@data-dita-seg-id]"):
-                seg_id = seg_elem.get("data-dita-seg-id")
-                xpath = skel_et.getpath(seg_elem)
-                src_match = src_tree.xpath(xpath)
-                tgt_match = tgt_tree.xpath(xpath)
-                if src_match and tgt_match:
-                    s_inner = utils.get_inner_xml(src_match[0]).strip()
-                    t_inner = utils.get_inner_xml(tgt_match[0]).strip()
-                    if s_inner == t_inner and s_inner:
-                        msg = f"Untranslated segment {seg_id} at {xpath}"
-                        warnings.append(msg)
-                        self.logger.warning(msg)
-
-        passed = not errors
-        for err in errors:
-            self.logger.error(err)
-        for warn in warnings:
-            self.logger.warning(warn)
-        self.logger.info("End validate")
-        return ValidationReport(passed, errors + warnings)
+        return self.validator.validate(src_xml, tgt_xml, skel_dir)
 
     def generate_dummy_translation(self, segments_json_path: str, output_path: str) -> str:
         segments_json_path = self._resolve(segments_json_path, self.intermediate_dir)
@@ -311,7 +246,7 @@ class Dita2LLM:
             Path to the generated translated XML and the validation report.
         """
         simple_xml_path = self._resolve(simple_xml_path, self.intermediate_dir)
-        self.logger.info(f"Start integrate from simple XML: {simple_xml_path}")
+        self.logger.info("Start integrate from simple XML: %s", simple_xml_path)
 
         name = os.path.splitext(os.path.basename(simple_xml_path))[0]
         if ".minimal" in name:
